Migrating docs to xml comments
