+++ src/Cake.Incubator/Cake.Incubator.csproj
@@ -55,6 +55,7 @@
     <Compile Include="DirectoryExtensions.cs" />
     <Compile Include="DotNetCoreTestExtensions.cs" />
     <Compile Include="EnvironmentExtensions.cs" />
+    <Compile Include="FileExtensions.cs" />
     <Compile Include="LoggingExtensions.cs" />
     <Compile Include="ProjectTypes.cs" />
     <Compile Include="StringExtensions.cs" />
